+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 - Fix #180: Use icon `fa-unlock` for public files
 - Fix #184: Display only published content files in the folder "Files from the stream"
 - Fix #186: Rename conflicted not published folder/file on creating/uploading
+- Fix #189: Initialize module content class
 
 0.16.1 - May 1, 2023
 --------------------